Introducing ClassFactory API to allow other components to define new classes according to their policy. This change maintains backward compatibility by providing DefaultClassFactory which is lazily initialized in ClassFile.defineInternal() method if ClassFactory was not provided via deprecated constructors. WildFly integration code will be changed to use jboss-module's ClassDefiner API and will integrate via this new API because it is not possible to access classloader non public methods and fields reflectively since JDK 12.
